Added pysimtest library for UWSim (#78)
* init branch

* created test xacro file

* Setup gate.xacro

* A simple test script file has been created which creates a .xml file from a
.xacro file with given params. The next step is to figure out which parts of
gate.xacro need to be abstracted out, and what range their values should be.
We should look into creating a script to generate sequences of parms which we
can then use to automate test parameter creation

* Arg generator has been created, as well as script which initializes the .xml
file for every combination of args. Next task is to run our robot's node so
that it communicates with the simulation, and create a C++ script to
read the simulation robot's pose and determine if it made it through the gate
in an alloted amount of time. We also need to determine which paramaters need
to be added to .xacro, and which values they can take

* Run Test Cases script now properly kills the processes needed to run
the simulation

* Cleaned up terminal output in run test cases

* Added test C++ scipt to emulate a ROS node determining test results, as well as
started the actual test_result_node.cpp which will determine test results based
on messages from /g500/pose

* Updated test script with simulation launch file, not sure if it is the right
one though. Also included functionality to assess what percent of tests passed

* added python test suite functionalities

* Reorganized simulation package structure

* added simTestBuilder in its own module

* cleanup some files

* implemented python interface for tests

* Added pysimtest library, needs testing

* Formatted pysimtest package

* Added main entry point

* Added main functionality

* recommit

* changed vehicle to seperate xacro

* merging into owen/gate

* fixed vehicle generated text issues

* comments for function remove_generated_xacro in file process.py

* Refactored runner and replaced .xacro functionality with .xml

* Documented and tested, 96% coverage

* Modified README.md, finished testing wiht 99% coverage

* merge with owen/gate

* stash pop

* fixed merge conflict (for real)

* after stash pop, adding changes to owen/gate

* final commit before pull request to master

* adding .pyc files to .gitignore

* after changing .gitignore

* deleted dynamics.py and all references to use_dynamics

* fixed issue with simulator ai

* removed all files associated with dynamics